Genesis file includes initial staking transactions
On startup the genesis.json file is read and all included transaction
are executed. Those transactions are applied to the state and the
initial AppHash and InitialValidatorSet is returned to Tendermint.
