+++ baseapp/baseapp.go
@@ -231,8 +231,10 @@ func (app *BaseApp) InitChain(req abci.RequestInitChain) (res abci.ResponseInitC
 
 	// Initialize the deliver state and run initChain
 	app.setDeliverState(abci.Header{})
-	app.initChainer(app.deliverState.ctx, req) // no error
+	res = app.initChainer(app.deliverState.ctx, req) // no error
 
+	// XXX-adrian: We want to commit here. To register the initial validator set
+	// initChainer needs return the initial validator set to Tendermint.
 	// NOTE: we don't commit, but BeginBlock for block 1
 	// starts from this deliverState
